extract BPartnerQuery to its own file
https://github.com/metasfresh/metasfresh/issues/5098
